+++ src/main/java/oripa/domain/fold/foldability/ring/RingElement.java
@@ -16,9 +16,9 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.domain.fold.foldability;
+package oripa.domain.fold.foldability.ring;
 
-class RingElement<V> {
+public class RingElement<V> {
 	private RingElement<V> next;
 	private RingElement<V> previous;
